[feat] : (배달) 리뷰 리스트 불러오기

+++ FoodDelivery/app/src/main/java/com/soten/fooddelivery/widget/adapter/viewholder/review/RestaurantReviewViewHolder.kt
@@ -31,10 +31,10 @@ class RestaurantReviewViewHolder(
             } else {
                 reviewThumbnailImage.isGone = true
             }
-
             reviewTitleText.text = model.title
             reviewText.text = model.description
-            ratingBar.rating = model.grade.toFloat()
+
+            ratingBar.rating = model.grade
         }
     }
